Migrate store files from .tsx to .ts and update file imports
