+++ src/store/reducers/device.reducer.ts
