+++ src/store/reducers/router.reducer.ts
@@ -1,19 +1,19 @@
-import { RouterAction, RouterActionTypes } from '../actions/router.actions';
+import { RouterAction, RouterActionTypes } from '../actions/router.action';
 import { Location } from 'react-router-dom';
 import { LocationState } from '../../core/models/router.model';
 import { Action } from 'history';
 
-export interface RouterState {
+export interface RouterModel {
   location: (Location & { state?: LocationState }) | null;
   action: string | null;
 }
 
-const initialState: RouterState = {
+const initialState: RouterModel = {
   location: null,
   action: null,
 };
 
-const RouterReducer = (state: RouterState = initialState, action: RouterAction) => {
+const RouterReducer = (state: RouterModel = initialState, action: RouterAction) => {
   switch (action.type) {
     case RouterActionTypes.LOCATION_CHANGE:
       return {
