+++ src/store/reducers/searching.reducer.ts
