+++ src/store/reducers/spinner.reducer.ts
