+++ src/store/reducers/store.reducer.ts
@@ -1,7 +1,7 @@
 import { convertPVPRankings, convertPVPTrain } from '../../core/pvp';
 import { replaceTempMovePvpName } from '../../util/utils';
 import { Store } from '../models/store.model';
-import { StoreActions } from '../../store/actions';
+import { StoreActions } from '../actions';
 import { StoreActionsUnion } from '../actions/store.action';
 import { IWeatherBoost } from '../../core/models/weatherBoost.model';
 import { isEqual } from '../../util/extension';
