+++ src/store/reducers/theme.reducer.ts
